Add files via upload
The Graphic Interface is very important for a code. If your Graphic Interface is not good enough, it probably will not be popular. Tkinter is the one of Graphic Interface in Python.
The advantages of Tkinter is easy to use and much simpler than other Graphic Interface designs, and Python is officially supporting Tkinter.